+++ csrc/id_model/schedule.cpp
@@ -5,7 +5,7 @@
  * SPDX-License-Identifier: BSD-3-Clause
  */
 // clang-format on
-#include <id_model/utils.h>
+#include <id_model/schedule.h>
 #include <val_graph.h>
 
 namespace nvfuser {
